% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area__of_rectangle.R
\name{area_of_rec}
\alias{area_of_rec}
\title{Calculate the area of a given rectangle}
\usage{
area_of_rec(len, wid)
}
\arguments{
\item{len}{Length of rectangle}

\item{wid}{Width of rectangle}
}
\value{
Returns the calculated area.
}
\description{
Calculate the area of a given rectangle
}
